Fix incorrect deny(missing_docs)
