+++ fog/overseer/server/src/lib.rs
@@ -1,7 +1,10 @@
 // Copyright (c) 2018-2022 The MobileCoin Foundation
 
+//! Fog overseer server library.
+
 #![feature(proc_macro_hygiene, decl_macro)]
-#[deny(missing_docs)]
+#![deny(missing_docs)]
+
 pub mod config;
 pub mod metrics;
 pub mod responses;
